Curso 2 - Aula 5

+++ src/main/kotlin/com/ajrdevops/alugames/modelo/Aluguel.kt
@@ -1,5 +1,6 @@
 package com.ajrdevops.alugames.modelo
 
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.Period
 
@@ -9,7 +10,7 @@ data class Aluguel(
     val periodo: Periodo
 )
 {
-    val valorAluguel:Double = gamer.plano.obterValor(this);
+    val valorAluguel:BigDecimal = gamer.plano.obterValor(this);
     override fun toString(): String {
         return "\n Aluguel do ${jogo.titulo} por ${gamer.nome} pelo valor de R$ ${valorAluguel.toString()}"
     }
