+++ src/main/kotlin/com/ajrdevops/alugames/modelo/Plano.kt
@@ -1,8 +1,10 @@
 package com.ajrdevops.alugames.modelo
 
+import java.math.BigDecimal
+
 //sealed : classe abstract que as herdeiras tem que estar no mesmo pacote
 sealed class Plano(val tipo:String) {
-    open fun obterValor(aluguel: Aluguel):Double {
-        return aluguel.jogo.preco * aluguel.periodo.emDias
+    open fun obterValor(aluguel: Aluguel):BigDecimal {
+        return aluguel.jogo.preco * aluguel.periodo.emDias.toBigDecimal()
     }
 }
