+++ src/main/kotlin/com/ajrdevops/alugames/utilitario/JogoExtension.kt
@@ -2,5 +2,5 @@ import com.ajrdevops.alugames.modelo.InfoJogoJson
 import com.ajrdevops.alugames.modelo.Jogo
 
 fun InfoJogoJson.criaJogo(): Jogo {
-    return Jogo(this.titulo, this.capa, this.preco, this.descricao)
+    return Jogo(this.titulo, this.capa, this.preco.toBigDecimal(), this.descricao)
 }
